requirements_final.txt
async-timeout == 4.0.3       
black==23.3.0      
blinker==1.6.2       
certifi==2023.5.7    
charset-normalizer==3.1.0       
click==8.1.7       
docker==6.1.3       
Flask==2.3.2       
Flask-SSE==1.0.0       
gevent==23.9.1      
GPUtil==1.4.0       
gunicorn==21.2.0      
idna==3.4         
importlib-metadata==6.8.0       
itsdangerous==2.1.2       
Jinja2==3.1.2       
kafka-python==2.0.2   
loguru==0.7.2     
MarkupSafe==2.1.3       
math3d==3.4.1       
numpy==1.24.4      
opencv-python==4.8.0.76    
packaging==23.1        
pandas==2.0.3       
Pillow==10.0.0      
pip==20.0.2      
pkg-resources==0.0.0       
psutil==5.9.5       
py-cpuinfo==9.0.0       
pymodbus==2.5.3       
pymongo==4.3.3       
pyserial==3.5         
python-dateutil==2.8.2       
pytz==2023.3.post1
redis==4.5.5       
requests==2.31.0      
setuptools==44.0.0      
six==1.16.0      
tzdata==2023.3      
urllib3==2.0.3       
urx==0.11.0      
websocket-client==1.5.3       
werkzeug==2.3.7       
zipp==3.17.0 
